+++ BumboApp/Bumbo.App.Web/Bumbo - Backup.App.Web.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -14,14 +14,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Models\Repositorys\" />
+    <ProjectReference Include="..\Bumbo.Data\Bumbo.Data.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Bumbo.Data\Bumbo.Data.csproj" />
-    <ProjectReference Include="..\Bumbo.Domain\Bumbo.Domain.csproj" />
+    <Folder Include="Models\Repositorys\" />
   </ItemGroup>
-
-  
 
 </Project>
