--- BumboApp/Bumbo.App.Web/Models/Enums/DepartmentEnum.cs
@@ -1,10 +1,0 @@
-﻿namespace Bumbo.App.Web.Models.Enums
-{
-    public enum DepartmentEnum
-    {
-        Shelf,
-        Kassa,
-        Vers
-
-    }
-}
