+++ BumboApp/Bumbo.App.Web/Program.cs
@@ -18,7 +18,7 @@ namespace Bumbo.App.Web
             
             builder.Services.AddTransient<Bumbo.Data.Interfaces.INormRepository, Bumbo.Data.SqlRepository.NormRepository>();
             builder.Services.AddTransient<Bumbo.Data.Interfaces.IForecastRepository, Bumbo.Data.SqlRepository.ForecastRepository>();
-            builder.Services.AddTransient<Bumbo.App.Web.Models.Services.IGenerateForecastService, Bumbo.App.Web.Models.Services.GenerateForecastService>();
+            builder.Services.AddTransient<Bumbo.Domain.Services.Forecast.IGenerateForecastService, Bumbo.Domain.Services.Forecast.GenerateForecastService>();
