+++ BumboApp/Bumbo.App.Web/Views/Employee/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@using Bumbo.App.Web.Models;
+﻿@using Bumbo.App.Web.Models.ViewModels.Employee;
 @model EmployeeModel
 
 @{
