+++ BumboApp/Bumbo.App.Web/Views/Forecast/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@using Bumbo.App.Web.Models.Models
+﻿@using Bumbo.Domain.Models;
 @using Bumbo.App.Web.Models.ViewModels.Forecast;
 @using Microsoft.IdentityModel.Tokens
