+++ BumboApp/Bumbo.Domain/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace Bumbo.Domain
+{
+    public class Class1
+    {
+
+    }
+}
