+++ BumboApp/Bumbo.Domain/Enums/DepartmentEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace Bumbo.Domain.Enums;
+
+public enum DepartmentEnum
+{
+    Shelf,
+    Kassa,
+    Vers
+
+}
