+++ BumboApp/BumboApp.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.11.35222.181
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bumbo.App.Web", "Bumbo.App.Web\Bumbo.App.Web.csproj", "{FB30F333-73DB-47B6-B454-F396FF0E5EB4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bumbo.Data", "Bumbo.Data\Bumbo.Data.csproj", "{C9065521-D0D0-4D3F-8CAE-0BA5BC103B24}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bumbo.Data", "Bumbo.Data\Bumbo.Data.csproj", "{C9065521-D0D0-4D3F-8CAE-0BA5BC103B24}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bumbo.Domain", "Bumbo.Domain\Bumbo.Domain.csproj", "{C98051E3-073E-4EA8-8AFB-473686FE8DE0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{C9065521-D0D0-4D3F-8CAE-0BA5BC103B24}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9065521-D0D0-4D3F-8CAE-0BA5BC103B24}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C9065521-D0D0-4D3F-8CAE-0BA5BC103B24}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C98051E3-073E-4EA8-8AFB-473686FE8DE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C98051E3-073E-4EA8-8AFB-473686FE8DE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C98051E3-073E-4EA8-8AFB-473686FE8DE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C98051E3-073E-4EA8-8AFB-473686FE8DE0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
